+++ reactworkplace/20200817.md
@@ -1,0 +1,30 @@
+20200817 
+
+ 
+
+[*SO: Convert Unix time with PowerShell*] 
+
+[*SO: Given two directory trees, how can I find out which files differ by content?*] 
+
+[*SO: Call Microsoft Graph API Using Azure AD 1.0 Endpoint Client Credentials Flow*] 
+
+ 
+
+[*Common Microsoft Resources in Azure Active Directory*] OMG 
+
+ 
+
+[*Pandoc a universal document converter*] 
+
+ 
+
+[*Microsoft identity platform access tokens*] 
+
+ 
+
+  [*SO: Convert Unix time with PowerShell*]: https://stackoverflow.com/questions/63442637/access-token-question-in-azure-functions-authentication
+  [*SO: Given two directory trees, how can I find out which files differ by content?*]: https://stackoverflow.com/questions/4997693/given-two-directory-trees-how-can-i-find-out-which-files-differ-by-content
+  [*SO: Call Microsoft Graph API Using Azure AD 1.0 Endpoint Client Credentials Flow*]: https://stackoverflow.com/questions/46101980/call-microsoft-graph-api-using-azure-ad-1-0-endpoint-client-credentials-flow
+  [*Common Microsoft Resources in Azure Active Directory*]: https://www.shawntabrizi.com/aad/common-microsoft-resources-azure-active-directory/
+  [*Pandoc a universal document converter*]: https://pandoc.org/
+  [*Microsoft identity platform access tokens*]: https://docs.microsoft.com/en-us/azure/active-directory/develop/access-tokens
