+++ reactworkplace/w3school/src/index.js
@@ -8,25 +8,6 @@ ReactDOM.render(
   document.getElementById('root')
 )
 
-// ES 6
-class Header {
-  constructor() {
-    this.color = "Red";
-  }
-
-  changeColor = () => {
-    document.getElementById("demo").innerHTML += JSON.stringify(this);
-  }
-}
-
-var myheader = new Header();
-
-// The window object calls the function:
-window.addEventListener("load", myheader.changeColor);
-
-// A button object calls the function:
-document.getElementById("btn").addEventListener("click", myheader.changeColor);
-
 // React Render HTML
 ReactDOM.render(<p>Find control and render.</p>, document.getElementById('reactRenderHtml'))
